% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{runStmGui}
\alias{runStmGui}
\title{Helper function to launch the stm shiny app}
\usage{
runStmGui()
}
\description{
\code{runStmGui} launches the shiny app to help with walking through the
creation of stm models
}

