\name{stmgui-package}
\alias{stmgui-package}
\alias{stmgui}
\docType{package}
\title{
\packageTitle{stmgui}
}
\description{
\packageDescription{stmgui}
}
\details{
stmgui is a simple Shiny application that makes it easy to work with the
stm text processing package. Simply run \code{runStmGui()} to launch the
application.

The application intends to make it easy to perform the end-to-end process for
creating an stm model with visualizations.
}
\author{
\packageAuthor{stmgui}

Maintainer: \packageMaintainer{stmgui}
}
\references{
See the \code{\link[stm]{stm}} documentation for more details and usage
}
\keyword{ stm, shiny }
